requirements.txt
fastapi==0.112.2
uvicorn==0.30.6
requests==2.32.3
python-dotenv==1.0.1
openai==1.51.0
google-generativeai==0.7.2
